optim: faster hook for x64 via inline asm
This saves a pop, push and mov. I should note that now unfortunately it seems that there is an extra unused function generated which I don't know how to have clang remove, but it barely impacts size anyway, plus it doesn't matter, this is still faster than the old way :3c.